requirements.txt
streamlit
langchain
langchain-core
langchain-huggingface
langchain-community
langgraph
pokebase
requests
Pillow
sentence-transformers
chromadb
img2vec-pytorch
python-dotenv